Move ObsoleteGlobalInspectionTests for QuickFixes into relevant files under RD.QuickFixes
